Interactive excerpts table

Creates an interactive datatable where users can filter excerpts by code. Automatically detects all code columns starting with "c_".

Usage

view_excerpts_table(data)

Arguments

data

A data.frame or tibble with one column called excerpt and multiple code columns starting with "c_".

Value

A DT datatable widget.

Examples

library(dplyr)
library(tidyr)
library(DT)

df <- tibble::tibble(
  excerpt = c(
    "I felt connected to peers.",
    "We should normalize conversations about MH.",
    "My teachers helped me belong.",
    "I am comfortable talking about suicide."
  ),
  c_belonging = c(TRUE, FALSE, TRUE, FALSE),
  c_destigmatization = c(FALSE, TRUE, FALSE, FALSE),
  c_suicide_comfort = c(FALSE, FALSE, FALSE, TRUE)
)

view_excerpts_table(df)